requirements.txt
streamlit
PyMuPDF
langchain
langchain-community 
sentence-transformers
faiss-cpu
llama-cpp-python
numpy